Refactor SurahCards to use SurahCardView component
