+++ lib/presentation/widgets/widgets.dart
@@ -2,4 +2,5 @@ export 'app_scaffold.dart';
 export 'app_snackbar.dart';
 export 'floating_bottom_bar.dart';
 export 'scaffold_base_linear_color.dart';
+export 'surah_card_view.dart';
 export 'widget_switcher.dart';
